# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients": {
      "post": {
        "operationId": "PatientsController_create",
        "summary": "Create a new patient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The patient has been successfully created."
          },
          "400": {
            "description": "Bad Request. The input data is invalid."
          },
          "401": {
            "description": "Unauthorized. User needs to be authenticated."
          }
        },
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "patients"
        ]
      },
      "get": {
        "operationId": "PatientsController_findAll",
        "summary": "Retrieve all patients",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of patients retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/active": {
      "patch": {
        "operationId": "PatientsController_changeActive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{id}": {
      "delete": {
        "operationId": "PatientsController_delete",
        "summary": "Delete a specific patient by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the patient to be deleted.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient deleted successfully."
          },
          "404": {
            "description": "Not Found. Patient with the specified ID does not exist."
          }
        },
        "tags": [
          "patients"
        ]
      },
      "get": {
        "operationId": "PatientsController_findOne",
        "summary": "Retrieve a specific patient by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the patient.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found. Patient with the specified ID does not exist."
          }
        },
        "tags": [
          "patients"
        ]
      }
    },
...
    "/offices": {
      "get": {
        "operationId": "OfficesController_findAllFilters",
        "summary": "Retrieve all offices with optional filters",
        "parameters": [
          {
            "name": "sortKey",
            "required": false,
            "in": "query",
            "description": "The key to sort the offices by.",
            "schema": {
              "example": "specialty",
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "The direction to sort the offices. Can be \"ASC\" or \"DESC\".",
            "schema": {
              "example": "ASC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "The page number for pagination.",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": true,
            "in": "query",
            "description": "The number of offices per page.",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of offices retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllOfficeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The input data is invalid."
          },
          "404": {
            "description": "Not Found. No offices match the provided filters."
          }
        },
        "tags": [
          "offices"
        ]
      },
      "post": {
        "operationId": "OfficesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfficeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "offices"
        ]
      },
      "patch": {
        "operationId": "OfficesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfficeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "offices"
        ]
      }
    },
...
    "/offices/{id}": {
      "get": {
        "operationId": "OfficesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the office.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Office retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Office"
                }
              }
            }
          },
          "404": {
            "description": "Not Found. Office with the specified ID does not exist."
          }
        },
        "tags": [
          "offices"
        ]
      }
    },
...
  },
  "info": {
    "title": "MEDICINE_ANALYZING_SERVER_API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The first name of the patient. Only alphabetic characters are allowed.",
            "example": "Алексей"
          },
          "surname": {
            "type": "string",
            "description": "The surname of the patient. Only alphabetic characters are allowed.",
            "example": "Пухальский"
          },
          "patronymic": {
            "type": "string",
            "description": "The patronymic of the patient. This field is optional and can include alphabetic characters only.",
            "example": "Александрович"
          },
          "gender": {
            "type": "string",
            "description": "The gender of the patient. Must be one of the predefined values in the Gender enum.",
            "enum": [
              "Мужской",
              "Женский"
            ],
            "example": "Мужской"
          },
          "ageCategory": {
            "type": "string",
            "description": "The age category of the patient. Must be one of the predefined values in the AgeCategory enum.",
            "enum": [
              "взрослый",
              "дети"
            ],
            "example": "взрослый"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "The date of birth of the patient in ISO 8601 format.",
            "example": "1990-05-15"
          }
        },
        "required": [
          "name",
          "surname",
          "gender",
          "ageCategory",
          "dateOfBirth"
        ]
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the patient. Must be a valid UUID.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          },
          "name": {
            "type": "string",
            "description": "The first name of the patient. Only alphabetic characters are allowed.",
            "example": "Иван"
          },
          "surname": {
            "type": "string",
            "description": "The surname of the patient. Only alphabetic characters are allowed.",
            "example": "Иванов"
          },
          "patronymic": {
            "type": "string",
            "description": "The patronymic of the patient. This field is optional and can include alphabetic characters only.",
            "example": "Иванович"
          },
          "gender": {
            "type": "string",
            "description": "The gender of the patient. Must be one of the predefined values in the Gender enum.",
            "enum": [
              "Мужской",
              "Женский"
            ],
            "example": "Мужской"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "The date of birth of the patient in ISO 8601 format.",
            "example": "1990-01-01"
          },
          "ageCategory": {
            "type": "string",
            "description": "The age category of the patient. Must be one of the predefined values in the AgeCategory enum.",
            "enum": [
              "взрослый",
              "дети"
            ],
            "example": "взрослый"
          }
        },
        "required": [
          "id",
          "name",
          "surname",
          "gender",
          "dateOfBirth",
          "ageCategory"
        ]
      },
      "ChangeActiveDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the entity. Must be a valid UUID.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          }
        },
        "required": [
          "id"
        ]
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the patient.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          },
          "active": {
            "type": "boolean",
            "description": "Indicates whether the patient is active.",
            "example": true
          },
          "dateOfBirth": {
            "type": "string",
            "description": "The date of birth of the patient in ISO format.",
            "example": "1985-06-15"
          },
          "fullName": {
            "type": "string",
            "description": "The full name of the patient.",
            "example": "Иванов Иван Иванович"
          }
        },
        "required": [
          "id",
          "active",
          "dateOfBirth",
          "fullName"
        ]
      },
      "PatientsResponseDto": {
        "type": "object",
        "properties": {
          "patients": {
            "description": "List of patients.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatientDto"
            }
          }
        },
        "required": [
          "patients"
        ]
      },
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the doctor.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          },
          "fullName": {
            "type": "string",
            "description": "The full name of the doctor.",
            "example": "Иванов Иван Иванович"
          },
          "typeOfShifts": {
            "type": "string",
            "description": "The type of shifts the doctor works.",
            "enum": [
              "Первая смена",
              "Вторая смена",
              "Полная смена"
            ]
          }
        },
        "required": [
          "id",
          "fullName",
          "typeOfShifts"
        ]
      },
      "OfficeResponseEntityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the office.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          },
          "number": {
            "type": "number",
            "description": "The number of the office.",
            "example": 101
          },
          "specialty": {
            "type": "string",
            "description": "The specialty of the office.",
            "enum": [
              "Общая практика",
              "Акушерство и гинекология",
              "Хирургия",
              "Стоматология",
              "Терапия",
              "Офтальмология",
              "Педиатрия",
              "Кардиология",
              "Неврология",
              "Эндокринология",
              "Дерматология",
              "Психиатрия"
            ]
          },
          "doctors": {
            "description": "List of doctors in the office.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoctorDto"
            }
          },
          "status": {
            "type": "string",
            "description": "The status of the office.",
            "enum": [
              "Полный состав",
              "Один сотрудник",
              "Нет сотрудников"
            ]
          }
        },
        "required": [
          "id",
          "number",
          "specialty",
          "doctors",
          "status"
        ]
      },
      "GetAllOfficeResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "The total number of offices available.",
            "example": 100
          },
          "items": {
            "description": "List of office entities.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfficeResponseEntityDto"
            }
          }
        },
        "required": [
          "total",
          "items"
        ]
      },
      "CreateOfficeDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "number",
            "description": "The office number. Must be a positive integer.",
            "example": 101
          },
          "specialty": {
            "type": "string",
            "description": "The specialty associated with the office. Must be one of the predefined values in the Specialty enum.",
            "enum": [
              "Общая практика",
              "Акушерство и гинекология",
              "Хирургия",
              "Стоматология",
              "Терапия",
              "Офтальмология",
              "Педиатрия",
              "Кардиология",
              "Неврология",
              "Эндокринология",
              "Дерматология",
              "Психиатрия"
            ],
            "example": "Стоматология"
          }
        },
        "required": [
          "number",
          "specialty"
        ]
      },
      "UpdateOfficeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the office. Must be a valid UUID.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          },
          "number": {
            "type": "number",
            "description": "The office number. Must be a positive integer.",
            "example": 101
          },
          "specialty": {
            "type": "string",
            "description": "The specialty associated with the office. Must be one of the predefined values in the Specialty enum.",
            "enum": [
              "Общая практика",
              "Акушерство и гинекология",
              "Хирургия",
              "Стоматология",
              "Терапия",
              "Офтальмология",
              "Педиатрия",
              "Кардиология",
              "Неврология",
              "Эндокринология",
              "Дерматология",
              "Психиатрия"
            ],
            "example": "Стоматология"
          }
        },
        "required": [
          "id",
          "number",
          "specialty"
        ]
      },
      "Office": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the office.",
            "example": "e5b33c5e-1b9f-4f73-9b5e-9b83e55a9c56"
          },
          "number": {
            "type": "number",
            "description": "The number of the office.",
            "example": 101
          },
          "specialty": {
            "type": "string",
            "description": "The specialty of the office.",
            "enum": [
              "Общая практика",
              "Акушерство и гинекология",
              "Хирургия",
              "Стоматология",
              "Терапия",
              "Офтальмология",
              "Педиатрия",
              "Кардиология",
              "Неврология",
              "Эндокринология",
              "Дерматология",
              "Психиатрия"
            ]
          },
          "doctors": {
            "description": "List of doctors associated with the office.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "The current status of the office.",
            "enum": [
              "Полный состав",
              "Один сотрудник",
              "Нет сотрудников"
            ],
            "default": "Нет сотрудников"
          }
        },
        "required": [
          "id",
          "number",
          "specialty",
          "doctors",
          "status"
        ]
      },
...
    }
  }
}